requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0
python-dotenv>=1.0.1

chromadb>=0.5.0
sentence-transformers>=3.0.1
torch>=2.2.0

pymupdf>=1.24.7
pytesseract>=0.3.10
Pillow>=10.3.0

numpy>=1.26.4
requests>=2.31.0
tqdm>=4.66.4
python-multipart>=0.0.9

langdetect>=1.0.9

rank-bm25>=0.2.2          # BM25 lexical search for hybrid retrieval
sse-starlette>=1.8.2      # Server-Sent Events (streaming answers)
scikit-learn>=1.5.0       # needed for some vector ops / metrics
tiktoken>=0.7.0           # OpenAI tokenizer (for chunk/token budgeting)
